complete recommend model

+++ frontend/lib/constants/constant_values.dart
@@ -5,6 +5,6 @@ int calculateSavedAmount(int sPrice, int offPrice) {
   return (sPrice - offPrice).toInt();
 }
 
-String calculateOffPercentage(int sPrice, int offPrice) {
+String calculateOffPercentage(double sPrice, double offPrice) {
   return (((sPrice - offPrice) / sPrice) * 100).round().toString();
 }
